Update the Haptik SDK version in Demo App (#20)
