--- app/src/main/java/ai/haptik/android/sample/app/SavedCardActivity.java
@@ -1,13 +1,0 @@
-package ai.haptik.android.sample.app;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class SavedCardActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_saved_card);
-    }
-}
