+++ app/src/main/java/ai/haptik/android/sample/app/Utils.java
@@ -2,18 +2,10 @@ package ai.haptik.android.sample.app;
 
 import ai.haptik.android.sdk.HaptikLib;
 import ai.haptik.android.sdk.InitData;
+import ai.haptik.android.sdk.picassohelper.PicassoApiFactory;
 import android.app.Application;
-import android.content.Context;
-import android.content.SharedPreferences;
 
 public class Utils {
-    static final String PREFS_KEY_HAPTIK_DATA_SYNC_ONCE = "prefs_key_haptik_data_sync_once";
-    static final String PREFERENCES_FILE_NAME = "HAPTIKLIB_DEMO_APP_PREFS";
-
-    static void setHaptikInitialDataSyncDone(Context context, boolean val) {
-        SharedPreferences sharedPreferences = context.getSharedPreferences(PREFERENCES_FILE_NAME, Context.MODE_PRIVATE);
-        sharedPreferences.edit().putBoolean(PREFS_KEY_HAPTIK_DATA_SYNC_ONCE, val).apply();
-    }
 
     public static InitData getHaptikInitData(Application application) {
         //Comment the setRunEnvironement for while releasing it to production
@@ -23,6 +15,7 @@ public class Utils {
             .debugEnabled(BuildConfig.DEBUG)
             .notificationSound(R.raw.notification_sound) // Optional - only use if you want any non-default sound for haptik notification
             .verifyUserService(new UserVerificationService()) //Add only if guest user verification is required
+            .imageLoadingService(PicassoApiFactory.getPicassoApi()) // This is mandatory, please refer Image Loading modules from docs.haptik.ai
             .build();
     }
 }
