--- app/src/main/res/layout/activity_client_signup.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ai.haptik.android.sdk.widget.HaptikSignupProgessView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/hpv_progress"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@android:color/white"/>
