--- app/src/main/res/layout/activity_inbox.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-  <ai.haptik.android.sdk.extensions.inbox.InboxView
-      android:id="@+id/inbox"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent">
-
-  </ai.haptik.android.sdk.extensions.inbox.InboxView>
-</LinearLayout>
