--- app/src/main/res/layout/activity_saved_card.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_saved_card"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@android:color/white"
-    tools:context="ai.haptik.android.sample.app.SavedCardActivity">
-
-  <ai.haptik.android.sdk.payment.UserSavedCardsView
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      android:layout_marginTop="16dp"/>
-
-</RelativeLayout>
