+++ build.gradle
@@ -16,6 +16,13 @@ allprojects {
         google()
         jcenter()
         // ADD MAVEN URL FOR Haptik, TIL & PhonePe SDK - Ask Haptik about it or look into Integration Document
+        maven {
+            url "http://artifactory.hellohaptik.com:8081/artifactory/libs-release-local"
+            credentials {
+                username 'ASK HAPTIK FOR USERNAME'
+                password 'ASK HAPTIK FOR PASSWORD'
+            }
+        }
     }
 }
