+++ gradle.properties
@@ -17,7 +17,8 @@ org.gradle.jvmargs=-Xmx4608M
 # org.gradle.parallel=true
 
 android_support_library_version=27.0.2
-google_play_services_client_library_version=15.0.1
+google_play_services_client_library_version=16.0.0
 retrofit_library_version=2.2.0
-firebase_core = 16.0.0
-firebase_messaging = 17.3.0
+firebase_core = 16.0.6
+firebase_messaging = 17.6.0
+playServicesBase = 16.1.0
